still doesn't work, but differently

+++ main.py
@@ -5,14 +5,18 @@ import nltk
 import pandas as pd
 import torch
 from nltk.tokenize import word_tokenize
-from torch import optim, nn
+from torch import optim, nn, Tensor, IntTensor
 from torch.nn.modules.loss import _Loss
 from torch.optim import Optimizer
 from torch.types import Device
 from torch.utils.data import DataLoader, random_split
 
 from task1.MyRnn import MyRnn
-from task1.process_data import process_data, create_vocabulary, get_ngrams, create_dataset
+from task1.process_data import process_data, create_vocabulary, create_dataset
+
+
+def does_file_exist(file: str) -> bool:
+    return Path(file).is_file()
 
 
 def train_my_rnn(model: MyRnn, train_dataloader: DataLoader, validation_dataloader: DataLoader,
@@ -28,29 +32,30 @@ def train_my_rnn(model: MyRnn, train_dataloader: DataLoader, validation_dataload
     for epoch in range(epochs):
         start = time.time()
         model.train()
-        batch_X: torch.Tensor
-        batch_y: torch.Tensor
+        batch_lengths: IntTensor
+        batch_data: Tensor
         batch_train_losses = []
         epoch_train_correct_predictions = 0
         epoch_train_total_predictions = 0
-        for batch_X, batch_y in train_dataloader:
-            # move the data to the selected device
-            batch_X, batch_y = batch_X.to(device), batch_y.to(device)
+        for batch_lengths, batch_data in train_dataloader:
             # zero the gradient each batch
             optimizer.zero_grad()
-            # go over all ngrams in the batch
-            outputs = model(batch_X)
+            # fix data for the model
+            batch_data = batch_data.to(device)
+            # get the predictions
+            outputs: Tensor
+            outputs = model((batch_lengths, batch_data), model.create_hidden(train_dataloader.batch_size, device))
             # calculate the loss
-            loss: torch.Tensor = loss_function(outputs, batch_y)
+            loss: Tensor = loss_function(outputs.transpose(1, 2), batch_data[:, 1:-1])
             batch_train_losses.append(loss.item())
             # propagate the loss backwards
-            b = loss.backward()
+            loss.backward()
             # update the weights
             optimizer.step()
             # add the correct predictions
-            epoch_train_correct_predictions += (outputs.argmax(dim=1) == batch_y).sum().item()
+            epoch_train_correct_predictions += (outputs.argmax(dim=2) == batch_data[:, 1:-1]).sum().item()
             # add the total predictions
-            epoch_train_total_predictions += len(batch_y)
+            epoch_train_total_predictions += train_dataloader.batch_size
 
         # log the epoch time
         epoch_time = time.time()
@@ -58,33 +63,35 @@ def train_my_rnn(model: MyRnn, train_dataloader: DataLoader, validation_dataload
         train_losses.append(sum(batch_train_losses) / len(batch_train_losses))
         # add the final accuracy in the epoch to the list
         train_accuracies.append(epoch_train_correct_predictions / epoch_train_total_predictions)
+        # begin results printo 
+        print(f"Epoch: {epoch + 1}")
+        print(f"Train Accuracy: {train_accuracies[-1]}, Train Loss: {train_losses[-1]}")
         # calculate the accuracy and loss on the validation set
         batch_validation_losses = []
         validation_correct_predictions = 0
         validation_total_predictions = 0
         model.eval()
         with torch.no_grad():
-            batch_X: torch.Tensor
-            batch_y: torch.Tensor
-            for batch_X, batch_y in validation_dataloader:
-                # move the data to the selected device
-                batch_X, batch_y = batch_X.to(device), batch_y.to(device)
-                # add the batch size to the total predictions
-                validation_total_predictions += len(batch_y)
+            for batch_lengths, batch_data in validation_dataloader:
+                # fix data for the model
+                batch_data = batch_data.to(device)
                 # get the predictions
-                outputs = model(batch_X)
+                outputs: Tensor
+                outputs = model((batch_lengths, batch_data), model.create_hidden(train_dataloader.batch_size, device))
                 # calculate the loss
-                loss: torch.Tensor = loss_function(outputs, batch_y)
+                loss: Tensor = loss_function(outputs.transpose(1, 2), batch_data[:, 1:-1])
+                batch_train_losses.append(loss.item())
+                # add the batch size to the total predictions
+                validation_total_predictions += validation_dataloader.batch_size
+                # add the loss to the list
                 batch_validation_losses.append(loss.item())
                 # add the correct predictions
-                validation_correct_predictions += (outputs.argmax(dim=1) == batch_y).sum().item()
+                validation_correct_predictions += (outputs.argmax(dim=2) == batch_data[:, 1:-1]).sum().item()
         # get the accuracy of the epoch and put it in the accuracies list
         validation_accuracies.append(validation_correct_predictions / validation_total_predictions)
         # get the loss of the epoch and put it in the losses list
         validation_losses.append(sum(batch_validation_losses) / len(batch_validation_losses))
         # print the results
-        print(f"Epoch: {epoch + 1}")
-        print(f"Train Accuracy: {train_accuracies[-1]}, Train Loss: {train_losses[-1]}")
         print(f"Validation Accuracy: {validation_accuracies[-1]}, Validation Loss: {validation_losses[-1]}")
         print(f"Time: {int(epoch_time - start)} (+{int(time.time() - epoch_time)}) seconds")
         print("--------------------------------------------------")
@@ -99,7 +106,7 @@ if __name__ == '__main__':
     # check if the files exist
     train_data_file = Path(f"{save_dir}/train.csv")
     test_data_file = Path(f"{save_dir}/test.csv")
-    found = train_data_file.is_file() and test_data_file.is_file() 
+    found = train_data_file.is_file() and test_data_file.is_file()
     if found:
         # read files
         df_train, df_test = pd.read_csv(train_data_file), pd.read_csv(test_data_file)
@@ -110,7 +117,8 @@ if __name__ == '__main__':
                                          tokenizer=word_tokenize)
     # create the vocabulary object
     print("Creating vocabulary")
-    vocabulary = create_vocabulary([df_train, df_test])
+    vocabulary = create_vocabulary([df_train, df_test], ["<PAD>", "<UNK>"])
+    print(f"vocabulary size: {len(vocabulary)}")
     print("Finished creating vocabulary")
     # transform the tokens in the dataframes to their numeric equivalents
     print("Transforming tokens to numeric")
@@ -122,21 +130,45 @@ if __name__ == '__main__':
         df_train.to_csv(f"{save_dir}/train.csv")
         df_test.to_csv(f"{save_dir}/test.csv")
     # create the dataset and loader
-    print("Creating ngrams")
+    print("Getting numeric tokens")
     numeric_train: list[list[int]] = df_train['numeric'].to_list()
     numeric_test: list[list[int]] = df_test['numeric'].to_list()
-    train_ngrams = get_ngrams(numeric_train, 2)
-    test_ngrams = get_ngrams(numeric_test, 2)
-    print("Finished creating ngrams")
+    print("Finished getting numeric tokens")
     # create the tensor datasets
-    batch_size = 64
-    dataset, train_max_length = create_dataset(train_ngrams)
-    torch.save(dataset, "dataset.pth")
-    test_dataset, test_max_length = create_dataset(test_ngrams)
-    torch.save(test_dataset, "test_dataset.pth")
-    train_dataset, validation_dataset = random_split(dataset, [0.8, 0.2])
-    torch.save(train_dataset, "train_dataset.pth")
-    torch.save(validation_dataset, "validation_dataset.pth")
+    batch_size = 16
+    if does_file_exist("test_dataset.pth"):
+        print("Loading test dataset")
+        test_dataset = torch.load("test_dataset.pth")
+        print("Finished loading test dataset")
+    else:
+        print("Creating test dataset")
+        test_dataset = create_dataset(numeric_test)
+        print("Finished creating test dataset")
+        torch.save(test_dataset, "test_dataset.pth")
+        print("Saved test dataset")
+    if does_file_exist("train_dataset.pth") and does_file_exist("validation_dataset.pth"):
+        print("Loading datasets")
+        train_dataset = torch.load("train_dataset.pth")
+        validation_dataset = torch.load("validation_dataset.pth")
+        print("Finished loading datasets")
+    else:
+        if does_file_exist("dataset.pth"):
+            print("Loading dataset")
+            dataset = torch.load("dataset.pth")
+            print("Finished loading dataset")
+        else:
+            print("Creating dataset")
+            dataset = create_dataset(numeric_train)
+            print("Finished creating dataset")
+            torch.save(dataset, "dataset.pth")
+            print("Saved dataset")
+        print("Splitting dataset")
+        train_dataset, validation_dataset = random_split(dataset, [0.8, 0.2], torch.Generator().manual_seed(42))
+        print("Finished splitting dataset")
+        print("Saving datasets")
+        torch.save(train_dataset, "train_dataset.pth")
+        torch.save(validation_dataset, "validation_dataset.pth")
+        print("Saved datasets")
     # delete the dataframes to save some memory
     del df_train
     del df_test
